MythCenter-wide: filename at bottom of tree window, more room for filenames in edit_metadata window.
Make room for filename at bottom of tree window.

Revert display of watchedstate in detailed display area of a video.
It's more helpful to show watchedstate in list of videos.

Give more space to display filenames in edit_metadata window.
The space was available, why not use it?

The space was there, why not use it?