ttvdb: Force German for 'Tatort' or 'Ein Fall für Zwei'
These two series only have useful metadata in German.
There didn't seem to be any other mechanism to tell mythbackend
to fetch German metadata for these two and English for everything else.
It would be nice if it were as simple as running mythbackend temporarily
with a different LANG=, but that didn't seem to work.