Fix regression caused by commit 9297955
qt6: Overhaul the MythTVMenuNodeTuple structure (part 2)
It broke video & dvd chapter sub-menus.

Before that commit, while playing videos or dvds with chapters, there
was a chapter sub-menu that listed the times of chapter marks in the
content. You could choose one of these chapter marks to jump to.
After that commit, the chapter sub-menus are now empty.

This is first step, adding as much code as possible from the bad commit
that doesn't trigger the regression.